+++ src/main/java/org/hosh/runtime/DefaultCommandWrapper.java
@@ -51,6 +51,7 @@ public class DefaultCommandWrapper<T> implements Command, InterpreterAware {
 
 	@Override
 	public ExitStatus run(List<String> args, Channel in, Channel out, Channel err) {
+		interpreter.injectDeps(commandWrapper);
 		T resource = commandWrapper.before(args, in, out, err);
 		try {
 			for (;;) {