+++ src/main/java/org/hosh/runtime/Interpreter.java
@@ -117,7 +117,7 @@ public class Interpreter {
 		return run(statement, new NullChannel(), out, err);
 	}
 
-	public ExitStatus run(Statement statement, Channel in, Channel out, Channel err) {
+	protected ExitStatus run(Statement statement, Channel in, Channel out, Channel err) {
 		Command command = statement.getCommand();
 		injectDeps(command);
 		List<String> resolvedArguments = resolveArguments(statement.getArguments());
@@ -133,7 +133,7 @@ public class Interpreter {
 		Thread.currentThread().setName(name);
 	}
 
-	private void injectDeps(Command command) {
+	protected void injectDeps(Command command) {
 		command.downCast(StateAware.class).ifPresent(cmd -> cmd.setState(state));
 		command.downCast(TerminalAware.class).ifPresent(cmd -> cmd.setTerminal(terminal));
 		command.downCast(LineReaderAware.class).ifPresent(cmd -> cmd.setLineReader(lineReader));