+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- introducing `withLock` wrapper
 - improved error messages by using command name prefix (#88)
 
 ## [0.0.32] - 2019-05-22